fix: moved env to config and various other minor refactoring
