+++ .env.ci.test
@@ -82,9 +82,3 @@ TWITTER_CLIENT_SECRET=
 TWITTER_REDIRECT_URL=http://localhost:80/auth/login/twitter/callback
 
 TELESCOPE_ENABLED=false
-
-#Bioschemas Properties.
-NMRXIV_PROVIDER=NFDI4Chem
-NMRXIV_PROVIDER_URL=https://www.nfdi4chem.de/
-MEASUREMENT_TECHNIQUE=http://purl.obolibrary.org/obo/CHMO_0000613
-
