+++ .env.example
@@ -92,11 +92,6 @@ ORCID_ENVIRONMENT=sandbox
 
 TELESCOPE_ENABLED=false
 
-#Bioschemas Properties.
-NMRXIV_PROVIDER=NFDI4Chem
-NMRXIV_PROVIDER_URL=https://www.nfdi4chem.de/
-MEASUREMENT_TECHNIQUE=http://purl.obolibrary.org/obo/CHMO_0000613
-
 #DATACITE Properties
 DOI_HOST=datacite
 DATACITE_USERNAME=
